Added a download attempt counter to Track.  This is reset every session .. but when it reaches 10 tries, the download thread stops trying to grab this track.
I found I had 1 track that was trying to get downloaded and it just kept trying and trying.  This stops that from happening, but it will try again next session.

Suggestion: Maybe we should keep track of how many times this happens.  If it happens so many times, we should just remove this track, and assume it has been lost.  Or notify the server so the DB can be updated?