Various bug fixes for download and auto-normalize.
- It considered all missing, broken, and non-downloaded tracks (of which I had
hundreds) to be unrated.  I changed it to count only valid tracks as unrated.

- The download code was trying to download 350 tracks simultaneously, which
caused it to immediately gobble up all my memory.  So, I changed it to limit
the number of simultaneous downloads to 4.  It picks tracks to download
randomly out of the pending ones.  So, now it works nicely if you have a
large number of pending tracks to download (as I do).

- The download code had a slight inefficiency in it where it was saving the
track database many times when once would have been sufficient.  I fixed
that.

- Can't rate as "this sux".